[serving] Download model while initialize multi-node cluster
